+++ tests/resources/stargazers.json
